test: reorganization of the 'utils' unit test
