+++ tests/unit/CMakeLists.txt
@@ -15,4 +15,4 @@ add_subdirectory(mr)
 add_subdirectory(peer)
 add_subdirectory(private_data)
 add_subdirectory(template)
-add_subdirectory(utils-test)
+add_subdirectory(utils)
